% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{extract_params}
\alias{extract_params}
\title{Extract the parameters out of an optimized model.}
\usage{
extract_params(cur_fit)
}
\arguments{
\item{cur_fit}{A CmdStanMLE object}
}
\description{
Extract the parameters out of an optimized model.
}
