% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{extract_sf}
\alias{extract_sf}
\title{Extract the parameters out of an optimized model.}
\usage{
extract_sf(cur_fit)
}
\arguments{
\item{cur_fit}{A CmdStanMLE object}

\item{stan_data}{The formatted Stan data used to fit the model.}
}
\description{
Extract the parameters out of an optimized model.
}
