% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{modify_design}
\alias{modify_design}
\title{Modify the design by adding an interaction by genes}
\usage{
modify_design(design_formula, feat_name)
}
\arguments{
\item{design_formula}{The design formula}

\item{feat_name}{The variable name for features}
}
\description{
Modify the design by adding an interaction by genes
}
