% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{split_formula}
\alias{split_formula}
\title{Split design formula into fixed and random effects.}
\usage{
split_formula(design_formula)
}
\arguments{
\item{design_formula}{The design formula}
}
\description{
Split design formula into fixed and random effects.
}
