% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{disize}
\alias{disize}
\title{Design-informed size factor estimation}
\usage{
disize(
  design_formula,
  counts,
  metadata,
  batch_name,
  obs_name = "obs_id",
  n_feats = min(10000L, ncol(counts)),
  n_subset = 50L,
  n_iters = 5000L,
  rel_tol = 10000,
  init_alpha = 1e-08,
  history_size = 10L,
  n_threads = 1L,
  n_retries = 2L,
  verbose = 3L
)
}
\arguments{
\item{design_formula}{The formula describing the experimental design.}

\item{counts}{A (observation x feature) count matrix.}

\item{metadata}{A dataframe containing observation-level metadata.}

\item{batch_name}{The identifier for the batch column in 'metadata'.}

\item{obs_name}{The identifier for the observation column in 'metadata'.}

\item{n_feats}{The number of features used during estimation.}

\item{n_subset}{The number of observations per experimental unit used during
estimation, defaults to 50 (useful for scRNA-seq experiments).}

\item{n_iters}{The number of iterations used for estimation.}

\item{rel_tol}{The relative tolerance used for convergence.}

\item{init_alpha}{The initial step-size for the optimizer, lower values
can sometimes make it easier to estimate size factors for more complex
designs.}

\item{history_size}{The number of past updates to use for the L-BFGS
algorithm.}

\item{n_threads}{The number of threads to use for parallel processing.}

\item{n_retries}{The maximum number of times to retry fitting.}

\item{verbose}{The verbosity level (\code{1}: only errors, \code{2}: also allows
warnings,\code{3}: also allows messages, \code{4}: also prints additional
output useful for debugging).}
}
\value{
A named numeric vector containing the size factor estimates.
}
\description{
Design-informed size factor estimation
}
