% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{disize}
\alias{disize}
\title{Design-informed size factor estimation}
\usage{
disize(
  design_formula,
  counts,
  metadata,
  batch_name = "batch_id",
  obs_name = "obs_id",
  n_feats = min(10000L, ncol(counts)),
  n_subset = 50L,
  n_iters = "auto",
  n_threads = max(1, parallel::detectCores() - 2L),
  grainsize = ceiling(n_feats/n_threads),
  init_alpha = 1e-06,
  verbose = 3L
)
}
\arguments{
\item{design_formula}{The formula describing the experimental design.}

\item{counts}{A (obsservation x feature) count matrix.}

\item{metadata}{A dataframe containing sample metadata.}

\item{batch_name}{The identifier for the batch column in 'metadata',
defaults to "batch_id".}

\item{obs_name}{The identifier for the observation column in 'metadata',
defaults to "obs_id".}

\item{n_feats}{The number of features used during estimation, defaults to
automatically choosing based on a heuristic.}

\item{n_subset}{The number of observations per experimental unit used during
estimation, defaults to 50.}

\item{n_iters}{The number of iterations used for a single optimization pass.}

\item{n_threads}{The number of threads to use for parallel processing.}

\item{grainsize}{The grainsize used for partitioning data across threads (
highly encouraged to leave to its default value).}

\item{verbose}{The verbosity level (\code{1}: only errors, \code{2}: also allows warnings,
\code{3}: also allows messages, \code{4}: also prints additional output useful for
debugging).}
}
\value{
A named numeric vector containing the size factor point estimates.
}
\description{
Design-informed size factor estimation
}
