% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{disize}
\alias{disize}
\title{Estimate size factors}
\usage{
disize(
  design_formula,
  counts = NULL,
  metadata = NULL,
  model_data = NULL,
  batch_name = "batch",
  n_genes = 500
)
}
\arguments{
\item{design_formula}{The formula describing the experimental design.}

\item{counts}{A samples x genes count matrix.}

\item{metadata}{A dataframe containing sample metadata.}

\item{model_data}{The model data if already formatted.}

\item{batch_name}{The identifier for the batch column in 'metadata'.}

\item{n_genes}{The number of genes used to estimate size factors.}
}
\description{
Estimate size factors
}
