% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{disize}
\alias{disize}
\title{Design-infored size factor estimation.}
\usage{
disize(
  design_formula,
  counts = NULL,
  metadata = NULL,
  model_data = NULL,
  batch_name = "batch",
  obs_name = "obs_id",
  feat_name = "feat_id",
  n_feats = 500,
  n_subset = 50,
  n_threads = NULL,
  verbose = 3,
  n_passes = 20,
  n_iters = 100,
  tolerance = 0.001
)
}
\arguments{
\item{design_formula}{The formula describing the experimental design.}

\item{counts}{A (obsservation x feature) count matrix.}

\item{metadata}{A dataframe containing sample metadata.}

\item{model_data}{The model data if already constructed. Must contain the
'batch_name' and "feat_idx" identifiers as columns, and any
predictors used in 'design_formula'.}

\item{batch_name}{The identifier for the batch column in 'metadata',
defaults to "batch".}

\item{n_feats}{The number of genes used during estimation, defaults to 500.
Increasing this value will result in this function taking longer but more
confidence in the size factors.}

\item{n_subset}{The number of observations per experimental unit used during
estimation, defaults to 50.}

\item{n_threads}{The number of threads to be used during estimation,
defaults to 1. Increasing this value will generally decrease runtime.}

\item{backend}{How to call Stan, defaults to "rstan".}
}
\value{
A named numeric vector containing the size factor point estimates.
}
\description{
Design-infored size factor estimation.
}
