% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{disize}
\alias{disize}
\title{Design-infored size factor estimation.}
\usage{
disize(
  design_formula,
  counts = NULL,
  metadata = NULL,
  model_data = NULL,
  batch_name = "batch",
  obs_name = "obs",
  gene_name = "gene",
  n_genes = 500,
  n_threads = 1
)
}
\arguments{
\item{design_formula}{The formula describing the experimental design.}

\item{counts}{A (obsservation x feature) count matrix.}

\item{metadata}{A dataframe containing sample metadata.}

\item{model_data}{The model data if already constructed. Must contain the
'batch_name', 'obs_name', 'gene_name' identifiers as columns and any
predictors used in 'design_formula'.}

\item{batch_name}{The identifier for the batch column in 'metadata',
defaults to "batch".}

\item{obs_name}{The identifier for the observation column in 'metadata',
defaults to "obs".}

\item{gene_name}{The identifier for the gene column in 'model_data',
defaults to "gene".}

\item{n_genes}{The number of genes used during estimation, defaults to 500.
Increasing this value will result in this function taking longer but more
confidence in the size factors.}

\item{n_threads}{The number of threads to be used during estimation,
defaults to 1. Increasing this value will generally decrease runtime.}
}
\value{
A named numeric vector containing the size factor point estimates.
}
\description{
Design-infored size factor estimation.
}
