% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{modify_design}
\alias{modify_design}
\title{Modify the design by adding an interaction by genes}
\usage{
modify_design(formula)
}
\arguments{
\item{formula}{this do stuff}
}
\description{
Modify the design by adding an interaction by genes
}
